requirements.txt
pymongo
mysql-connector-python
pymupdf